Change int for size_t in AES

+++ BlangDecrypt.cpp
@@ -87,10 +87,10 @@ std::byte *HashData(const std::byte *data1, const size_t data1Len, const std::by
  * @param ciphertext Buffer to store the ciphertext in
  * @return Length of the ciphertext
  */
-int32_t EncryptData(const unsigned char *plaintext, const size_t plaintextLen, const unsigned char *key, const unsigned char *iv, unsigned char *ciphertext)
+size_t EncryptData(const unsigned char *plaintext, const size_t plaintextLen, const unsigned char *key, const unsigned char *iv, unsigned char *ciphertext)
 {
     int32_t len;
-    int32_t ciphertextLen;
+    size_t ciphertextLen;
 
     EVP_CIPHER_CTX *ctx = EVP_CIPHER_CTX_new();
     EVP_EncryptInit_ex(ctx, EVP_aes_128_cbc(), nullptr, key, iv);
@@ -117,10 +117,10 @@ int32_t EncryptData(const unsigned char *plaintext, const size_t plaintextLen, c
  * @param plaintext Buffer to store the plaintext in
  * @return Length of the plaintext
  */
-int32_t DecryptData(const unsigned char *ciphertext, const size_t ciphertextLen, const unsigned char *key, const unsigned char *iv, unsigned char *plaintext)
+size_t DecryptData(const unsigned char *ciphertext, const size_t ciphertextLen, const unsigned char *key, const unsigned char *iv, unsigned char *plaintext)
 {
     int32_t len;
-    int32_t plaintextLen;
+    size_t plaintextLen;
 
     EVP_CIPHER_CTX *ctx = EVP_CIPHER_CTX_new();
     EVP_DecryptInit_ex(ctx, EVP_aes_128_cbc(), nullptr, key, iv);
@@ -150,7 +150,7 @@ int32_t DecryptData(const unsigned char *ciphertext, const size_t ciphertextLen,
 std::vector<std::byte> CryptData(const bool decrypt, const std::byte *inputData, const size_t inputDataLen, const std::byte *key, const std::byte *iv)
 {
     unsigned char *output = new unsigned char[inputDataLen + 16 - (inputDataLen % 16)];
-    uint64_t newSize;
+    size_t newSize;
 
     if (decrypt) {
         newSize = DecryptData((unsigned char*)inputData, inputDataLen, (unsigned char*)key, (unsigned char*)iv, output);
